DatabaseCapacity2 Removal
(1) Remove DatabaseCapacity2 from program (2) Move DataCapacity2 to
Storage (3) Add DatabaseCpacity2 logic into DatabaseCapacity1
